changed goal progress field

+++ core/models.py
@@ -9,7 +9,7 @@ class Goal(models.Model):
     description = models.TextField()
     prize = models.TextField()
     date_created = models.DateTimeField(auto_now_add=True)
-    progress_value = models.IntegerField()
+    progress_value = models.FloatField()
     
 class Participant(models.Model):
     user = models.OneToOneField(User)
